Fixed more warnings. Used vm_address_t in one case instead of mach_vm_address_t.. What a mistake.

+++ ZGCalculator.m
@@ -166,7 +166,7 @@
 					}
 					
 					mach_vm_size_t size = process->is64Bit ? sizeof(int64_t) : sizeof(int32_t);
-					void *value = malloc(size);
+					void *value = malloc((size_t)size);
 					
 					NSMutableString *newExpression;
